+++ hack/tests/test-subcommand.sh
@@ -8,7 +8,7 @@ pushd test/test-framework
 operator-sdk test local ./test/e2e
 
 # test operator-sdk test flags
-operator-sdk test local ./test/e2e --global-manifest deploy/crds/cache_memcacheds_crd.yaml --namespaced-manifest deploy/namespace-init.yaml --go-test-flags "-parallel 1" --kubeconfig $HOME/.kube/config --image=quay.io/coreos/operator-sdk-dev:test-framework-operator-runtime
+operator-sdk test local ./test/e2e --global-manifest deploy/crds/cache.example.com_memcacheds.yaml --namespaced-manifest deploy/namespace-init.yaml --go-test-flags "-parallel 1" --kubeconfig $HOME/.kube/config --image=quay.io/coreos/operator-sdk-dev:test-framework-operator-runtime
 
 # we use the test-memcached namespace for all future tests, so we only need to set this trap once
 kubectl create namespace test-memcached
@@ -28,12 +28,12 @@ kubectl delete namespace test-memcached
 
 # test operator in no-setup mode
 kubectl create namespace test-memcached
-kubectl create -f deploy/crds/cache_memcacheds_crd.yaml
+kubectl create -f deploy/crds/cache.example.com_memcacheds.yaml
 # this runs after the popd at the end, so it needs the path from the project root
-trap_add 'kubectl delete -f test/test-framework/deploy/crds/cache_memcacheds_crd.yaml' EXIT
-kubectl create -f deploy/crds/cache_memcachedrs_crd.yaml
+trap_add 'kubectl delete -f test/test-framework/deploy/crds/cache.example.com_memcacheds.yaml' EXIT
+kubectl create -f deploy/crds/cache.example.com_memcachedrs.yaml
 # this runs after the popd at the end, so it needs the path from the project root
-trap_add 'kubectl delete -f test/test-framework/deploy/crds/cache_memcachedrs_crd.yaml' EXIT
+trap_add 'kubectl delete -f test/test-framework/deploy/crds/cache.example.com_memcachedrs.yaml' EXIT
 kubectl create -f deploy/service_account.yaml --namespace test-memcached
 kubectl create -f deploy/role.yaml --namespace test-memcached
 kubectl create -f deploy/role_binding.yaml --namespace test-memcached